+++ src/fragmentmnp/fragmentmnp.py
@@ -59,14 +59,16 @@ class FragmentMNP():
                                 self.psd,
                                 self.data['fsd_beta'])
         self.density = data['density']
-        # Stop Pylance complaining about k_frag and k_diss not being present
+        # Stop Pylance complaining about k_frag, k_diss and k_min not being
+        # present (or being the wrong type) by declaring them as NumPy arrays
         self.k_frag = np.empty((self.n_size_classes, self.n_timesteps))
         self.k_diss = np.empty((self.n_size_classes, self.n_timesteps))
+        self.k_min = np.empty((self.n_timesteps,))
         # Calculate the rate constant distributions. If we've been given
         # a dict in data, use the contained params to create the distribution.
         # Else, presume that we've been given a scalar (validation will make
         # sure this is so) and use that as the average.
-        for k in ['k_frag', 'k_diss']:
+        for k in ['k_frag', 'k_diss', 'k_min']:
             if isinstance(data[k], dict):
                 k_f = data[k]['k_f']
                 k_0 = data[k]['k_0']
@@ -79,17 +81,24 @@ class FragmentMNP():
                 k_0 = 0.0
                 is_compound = True
                 params = {}
-            # Calculate the 2D (s, t) distribution
-            k_dist = self.set_k_distribution(dims={'s': self.surface_areas,
-                                                   't': self.t_grid},
-                                             k_f=k_f, k_0=k_0,
-                                             params=params,
-                                             is_compound=is_compound)
-            # If the rate constant is k_frag, then no fragmentation is
-            # allowed from the smallest size class and therefore we
-            # manually set this to zero
-            if k == 'k_frag':
-                k_dist[0, :] = 0.0
+            # If k_frag or k_diss, we want to calculate a 2D (s, t) distribution,
+            # and if k_min, we just want a 1D (t) distribution
+            if k in ['k_frag', 'k_diss']:
+                k_dist = self.set_k_distribution(dims={'s': self.surface_areas,
+                                                       't': self.t_grid},
+                                                 k_f=k_f, k_0=k_0,
+                                                 params=params,
+                                                 is_compound=is_compound)
+                # If the rate constant is k_frag, then no fragmentation is
+                # allowed from the smallest size class and therefore we
+                # manually set this to zero
+                if k == 'k_frag':
+                    k_dist[0, :] = 0.0
+            else:
+                k_dist = self.set_k_distribution(dims={'t': self.t_grid},
+                                                 k_f=k_f, k_0=k_0,
+                                                 params=params,
+                                                 is_compound=is_compound)
             # Check no values are less than zero
             if np.any(k_dist < 0.0):
                 msg = (f'Value for {k} distribution calculated from input '
@@ -111,7 +120,7 @@ class FragmentMNP():
 
         Notes
         -----
-        The model numerically solves the following differential
+        Internally the model numerically solves the following differential
         equation for each size class, to give a time series of mass
         concentrations `c`. `k` is the current size class, `i` are the
         daughter size classes.
@@ -129,64 +138,98 @@ class FragmentMNP():
         Mass concentrations are converted to particle number concentrations by
         assuming spherical particles with the density given in the input data.
         """
-
         # Define the initial value problem to pass to SciPy to solve.
         # This must satisfy c'(t) = f(t, c) with initial values given in data.
-        def f(t, c):
+        def f(t, c_big):
+            """
+            c_big has length N+1:
+              c_big[:N]   = mass in each of N size classes
+              c_big[N]    = total dissolved mass
+            """
             # Get the number of size classes and create results to be filled
             N = self.n_size_classes
-            dcdt = np.empty(N)
+            # Unpack
+            c_particles = c_big[:N]
+            c_dissolved = c_big[N]
+            c_min = c_big[N + 1]   # Not used as a source in any ODE (only grows)
             # Interpolate the time-dependent parameters to the specific
             # timestep given (which will be a float, rather than integer index)
             f_frag = interpolate.interp1d(self.t_grid, self.k_frag, axis=1,
                                           fill_value='extrapolate')
             f_diss = interpolate.interp1d(self.t_grid, self.k_diss, axis=1,
                                           fill_value='extrapolate')
+            f_min = interpolate.interp1d(self.t_grid, self.k_min, axis=0,
+                                         fill_value='extrapolate')
             k_frag = f_frag(t)
             k_diss = f_diss(t)
+            k_min = f_min(t)
+            # Build array of d/dt
+            dcdt = np.empty(N+2)
             # Loop over the size classes and perform the calculation
-            for k in np.arange(N):
+            for k in range(N):
                 # The differential equation that is being solved
-                dcdt[k] = - k_frag[k] * c[k] \
-                    + np.sum(self.fsd[:, k] * k_frag * c[:N]) \
-                    - k_diss[k] * c[k]
-            # Return the solution for all of the size classes
+                dcdt[k] = (
+                    - k_frag[k] * c_particles[k]
+                    + np.sum(self.fsd[:, k] * k_frag * c_particles[:N])
+                    - k_diss[k] * c_particles[k]
+                )
+
+            # Dissolved mass ODE:
+            #   Gains: sum of kdiss[k] * c_particles[k]
+            #   Loss:  k_min * c_dissolved
+            dcdt_dissolved = np.sum(k_diss * c_particles) - k_min * c_dissolved
+            # Assign to last entry
+            dcdt[N] = dcdt_dissolved
+
+            # Mineralized ODE:
+            #   Gains: k_min_avg * c_dissolved
+            #   No loss, so it's purely accumulative
+            dcdt_min = k_min * c_dissolved
+            dcdt[N + 1] = dcdt_min
+
             return dcdt
 
-        # Numerically solve this given the initial values for c
-        soln = solve_ivp(fun=f,
-                         method=self.config['solver_method'],
-                         t_span=(self.t_grid.min(), self.t_grid.max()),
-                         y0=self.initial_concs,
-                         t_eval=self.t_eval,
-                         rtol=self.config['solver_rtol'],
-                         atol=self.config['solver_atol'],
-                         max_step=self.config['solver_max_step'])
-        # If we didn't find a solution, raise an error
-        if not soln.success:
-            raise FMNPNumericalError('Model solution could not be ' +
-                                     f'found: {soln.message}')
-        # Calculate the timeseries of mass concentration lost to dissolution
-        # from the solution, first interpolating k_diss to the t values
-        # we evaluated the solution over (t_eval)
-        f_diss = interpolate.interp1d(self.t_grid, self.k_diss, axis=1,
-                                      fill_value='extrapolate')
-        k_diss_eval = f_diss(self.t_eval)
-        j_diss = k_diss_eval * soln.y
-        # Use this to calculate the cumulative mass concentration
-        # lost to dissolution
-        c_diss_from_sc = np.cumsum(j_diss, axis=1)
-        # Add initial concentration and sum across size classes
-        c_diss = np.sum(np.cumsum(j_diss, axis=1), axis=0) \
-            + self.initial_concs_diss
-        # Now we have the solutions as mass concentrations, we can
-        # convert to particle number concentrations
-        n = self.mass_to_particle_number(soln.y)
-        n_diss_from_sc = self.mass_to_particle_number(c_diss)
+        # Build the new N+12 initial conditions: N particulate states,
+        # +1 dissolved, +1 mineralized
+        y0 = np.concatenate([
+           self.initial_concs,          # microplastic mass per size class
+           [self.initial_concs_diss],   # dissolved mass
+           [0.0]                        # mineralized (initially zero)
+        ])
 
-        # Return the solution in an FMNPOutput object
-        return FMNPOutput(soln.t, soln.y, n, c_diss_from_sc, c_diss,
-                          n_diss_from_sc, soln, self.psd)
+        # Solve ODE
+        soln = solve_ivp(
+            fun=f,
+            method=self.config['solver_method'],
+            t_span=(self.t_grid.min(), self.t_grid.max()),
+            y0=y0,
+            t_eval=self.t_eval,
+            rtol=self.config['solver_rtol'],
+            atol=self.config['solver_atol'],
+            max_step=self.config['solver_max_step']
+        )
+        if not soln.success:
+            raise FMNPNumericalError('Model solution could not be found: ' +
+                                     f'{soln.message}')
+
+        # Extract solution
+        c_part_sol = soln.y[:self.n_size_classes, :]
+        c_diss_sol = soln.y[self.n_size_classes, :]
+        c_min_sol = soln.y[self.n_size_classes + 1, :]
+
+        # Convert microparticle mass to particle number
+        n_part_sol = self.mass_to_particle_number(c_part_sol)
+
+        # Build your FMNPOutput object in whatever format is required
+        return FMNPOutput(
+           soln.t,
+           c_part_sol,
+           n_part_sol,
+           c_diss_sol,
+           c_min_sol,
+           soln,
+           self.psd,
+        )
 
     def mass_to_particle_number(self, mass):
         """
@@ -220,13 +263,18 @@ class FragmentMNP():
         r"""
         Create a distribution based on the rate constant scaling factor ``k_f``
         and baseline adjustment factor ``k_0``. The distribution will be a
-        compound combination of power law / polynomial, exponential,
-        logarithmic and logistic regressions, encapsulated in the function
-        :math:`X(x)`, and have dimensions given by `dims`. For a distribution
-        with `D` dimensions:
+        compound or additive combination of power law / polynomial,
+        exponential, logarithmic and logistic regressions, encapsulated in the
+        function :math:`X(x)`, and have dimensions given by `dims`. For a
+        distribution with `D` dimensions:
 
         .. math::
             k(\mathbf{x}) = k_f \prod_{d=1}^D X(x_d) + k_0
+
+        or
+
+        .. math::
+            k(\mathbf{x}) = k_f \sum_{d=1}^D X(x_d) + k_0
 
         :math:`X(x)` is then given either by: